refactor to create menu component
create menu component
