+++ src/tests/providers/themeprovider/themeProvider.test.jsx
@@ -2,8 +2,10 @@ import React, { useContext } from 'react';
 import { HashRouter } from 'react-router-dom';
 import { render, within, fireEvent, screen } from '@testing-library/react';
 import { act } from 'react-dom/test-utils';
-import ThemeProvider, { ThemeContext } from '../../../providers/Theme';
-import AuthProvider from '../../../providers/Auth';
+import ThemeProvider, {
+  ThemeContext,
+} from '../../../providers/Theme/LocalTheme.provider';
+import AuthProvider from '../../../providers/Auth/Auth.provider';
 
 describe('local theme provider', () => {
   it('sets initialy dark theme', () => {
